style: made toggleGroupField and made PetToggle and ServiceToggle based on it
